story with location

+++ app/src/main/java/me/varoa/sad/core/data/remote/json/StoryJson.kt
@@ -9,6 +9,6 @@ data class StoryJson(
     val description: String,
     val photoUrl: String,
     val createdAt: String,
-    val lat: Float? = null,
-    val lon: Float? = null
+    val lat: Double? = null,
+    val lon: Double? = null
 )
